Remove implementation code

+++ persistence/ibc.go
@@ -14,14 +14,14 @@ import (
 // SetIBCStoreEntry sets the key value pair in the IBC store postgres table at the current height
 func (p *PostgresContext) SetIBCStoreEntry(key, value []byte) error {
 	ctx, tx := p.getCtxAndTx()
-	if _, err := tx.Exec(ctx, pTypes.InsertIBCStoreEntryQuery(uint64(p.Height), key, value)); err != nil {
+	if _, err := tx.Exec(ctx, pTypes.InsertIBCStoreEntryQuery(p.Height, key, value)); err != nil {
 		return err
 	}
 	return nil
 }
 
 // GetIBCStoreEntry returns the stored value for the key at the height provided from the IBC store table
-func (p *PostgresContext) GetIBCStoreEntry(key []byte, height uint64) ([]byte, error) {
+func (p *PostgresContext) GetIBCStoreEntry(key []byte, height int64) ([]byte, error) {
 	ctx, tx := p.getCtxAndTx()
 	row := tx.QueryRow(ctx, pTypes.GetIBCStoreEntryQuery(height, key))
 	var valueHex string
@@ -50,7 +50,7 @@ func (p *PostgresContext) SetIBCEvent(event *coreTypes.IBCEvent) error {
 		return err
 	}
 	eventHex := hex.EncodeToString(eventBz)
-	if _, err := tx.Exec(ctx, pTypes.InsertIBCEventQuery(uint64(p.Height), typeStr, eventHex)); err != nil {
+	if _, err := tx.Exec(ctx, pTypes.InsertIBCEventQuery(p.Height, typeStr, eventHex)); err != nil {
 		return err
 	}
 	return nil
